requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
pymysql
python-dotenv
